refactor: split tafsir panel components
