+++ app/(features)/surah/[surahId]/components/tafsir-panel/index.ts
@@ -1,3 +1,3 @@
-export { TafsirPanel } from '../TafsirPanel';
+export { TafsirPanel } from './TafsirPanel';
 export { useTafsirPanel } from './useTafsirPanel';
 export * from './tafsirPanel.utils';
