+++ app/(features)/tafsir/README.md
@@ -4,6 +4,6 @@ Shared components and utilities:
 
 - [SurahListSidebar](../../shared/SurahListSidebar.tsx) – navigation supplied by the features layout.
 - [AudioProvider](../../shared/player/context/AudioContext.tsx) – wraps layout for audio state.
-- [SettingsSidebar](../surah/[surahId]/components/SettingsSidebar.tsx) and [TafsirPanel](../surah/[surahId]/components/TafsirPanel.tsx) – reused UI panels from the Surah feature.
+- [SettingsSidebar](../surah/[surahId]/components/SettingsSidebar.tsx) and [TafsirPanel](../surah/[surahId]/components/tafsir-panel/TafsirPanel.tsx) – reused UI panels from the Surah feature.
 - [SettingsContext](../../providers/SettingsContext.tsx) – provides user translation and word settings.
 - [getVersesByChapter](../../../lib/api/verses.ts) and [getTafsirByVerse](../../../lib/api/tafsir.ts) – API helpers to load verse text and tafsir content.
